+++ src/client/user/User.ts
@@ -1,4 +1,6 @@
+import { AssetManager } from "@client/assets";
 import { Player } from "./player";
+import * as Babylon from "babylonjs";
 
 /**
  * The user currently associated with this session. An
@@ -12,14 +14,15 @@ export class User {
    * Create the player, the in-game component of this user. The
    * player should only exist if the world is loaded.
    */
-  spawnPlayer(): void {
-
+  async spawnPlayer(assetManager: AssetManager, scene: Babylon.Scene, canvas: HTMLCanvasElement) {
+    this.player = await Player.create(assetManager, scene, canvas);
   }
 
   /**
    * Remove the player, the in-game component of this user.
    */
   removePlayer(): void {
-
+    this.player?.dispose();
+    this.player = null;
   }
 }