+++ changelog.md
@@ -8,6 +8,7 @@
 * Rename to Ravensquest
 * Add in-game API calls to create in-game IOT w/ Venmo & Auth0 system
 * Ragdolls, Drones & Robots, Planes, Fun stuff
+* Add action queue
 
 ## 0.1.2